limpieza de details e incorporacion de Size a pasarela de compras
